Refactoring stars out of planets
(cherry picked from commit 1100ee58fc5c636dcd6d3505fc4e8c5c3368fe9a)
